refactor: inner dto class 분할 (PartnerDTO, VendorDTO)
