+++ src/main/java/io/saim/dash/coupon/issue/dto/IssueCreateRequestDTO.java
@@ -2,6 +2,9 @@ package io.saim.dash.coupon.issue.dto;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import io.saim.dash.coupon.common.dto.PartnerDTO;
+import io.saim.dash.coupon.common.dto.VendorDTO;
 import jakarta.annotation.Nullable;
 
 import lombok.AllArgsConstructor;
@@ -9,44 +12,7 @@ import lombok.Getter;
 
 @Getter @AllArgsConstructor
 public class IssueCreateRequestDTO {
-	private VendorRequestDTO vendor;
-	private PartnerRequestDTO partner;
+	private VendorDTO vendor;
+	private PartnerDTO partner;
 	private List<Long> products = new ArrayList<>();
-
-	@AllArgsConstructor
-	public static class VendorRequestDTO {
-		private String vendor_name;
-
-		@Nullable
-		private String president_name;
-
-		@Nullable
-		private String president_phone;
-
-		public String getVendorName() {
-			return vendor_name;
-		}
-
-		public String getPresidentName() {
-			return president_name;
-		}
-
-		public String getPresidentPhone() {
-			return president_phone;
-		}
-	}
-
-	@AllArgsConstructor
-	public static class PartnerRequestDTO {
-		private String business_name;
-		private String owner_phone;
-
-		public String getBusinessName() {
-			return business_name;
-		}
-
-		public String getOwnerPhone() {
-			return owner_phone;
-		}
-	}
 }
